move comment parsing codes from `comment_view.rs` to `client/parser.rs`
