+++ hackernews_tui/Cargo.toml
@@ -36,3 +36,4 @@ url = "2.2.2"
 config_parser2 = { version = "0.1.2", path = "../config_parser" }
 crossbeam-channel = "0.5.1"
 parking_lot = "0.11.2"
+lazy_static = "1.4.0"
