+++ hackernews_tui/src/view/comment_view.rs
@@ -9,52 +9,10 @@ use regex::Regex;
 
 type CommentComponent = HideableView<PaddedView<text_view::TextView>>;
 
-#[derive(Debug, Clone)]
-/// CommentState represents the state of a single comment component
-enum CommentState {
-    Collapsed,
-    PartiallyCollapsed,
-    Normal,
-}
-
-#[derive(Debug, Clone)]
-pub struct Comment {
-    state: CommentState,
-    top_comment_id: u32,
-    id: u32,
-
-    text: StyledString,
-    minimized_text: StyledString,
-
-    height: usize,
-    links: Vec<String>,
-}
-
-impl Comment {
-    pub fn new(
-        top_comment_id: u32,
-        id: u32,
-        text: StyledString,
-        minimized_text: StyledString,
-        height: usize,
-        links: Vec<String>,
-    ) -> Self {
-        Comment {
-            state: CommentState::Normal,
-            top_comment_id,
-            id,
-            text,
-            minimized_text,
-            height,
-            links,
-        }
-    }
-}
-
 /// CommentView is a View displaying a list of comments in a HN story
 pub struct CommentView {
     view: ScrollListView,
-    comments: Arc<RwLock<Vec<Comment>>>,
+    comments: Arc<RwLock<Vec<client::Comment>>>,
 
     raw_command: String,
 }
@@ -84,138 +42,6 @@ impl CommentView {
             raw_command: String::new(),
         };
         comment_view
-    }
-
-    fn decode_html(s: &str) -> String {
-        htmlescape::decode_html(s).unwrap_or_else(|_| s.to_string())
-    }
-
-    /// Parse a comment in HTML text style to markdown text style (with colors)
-    fn parse_single_comment(
-        s: &str,
-        paragraph_re: &Regex,
-        italic_re: &Regex,
-        code_re: &Regex,
-        link_re: &Regex,
-    ) -> (StyledString, Vec<String>) {
-        let mut s = paragraph_re.replace_all(s, "${paragraph}\n\n").to_string();
-        if s.ends_with("\n\n") {
-            s.remove(s.len() - 1);
-        }
-
-        s = italic_re.replace_all(&s, "*${text}*").to_string();
-        s = code_re.replace_all(&s, "```\n${code}\n```").to_string();
-
-        let mut links: Vec<String> = vec![];
-        let mut styled_s = StyledString::new();
-        // replace the `<a href="${link}">...</a>` pattern one-by-one with "${link}".
-        // cannot use `replace_all` because we want to replace a match string with a `StyledString`
-        loop {
-            match link_re.captures(&s.clone()) {
-                None => break,
-                Some(c) => {
-                    let m = c.get(0).unwrap();
-                    let link = Self::decode_html(c.name("link").unwrap().as_str());
-
-                    let range = m.range();
-                    let mut prefix: String = s
-                        .drain(std::ops::Range {
-                            start: 0,
-                            end: m.end(),
-                        })
-                        .collect();
-                    prefix.drain(range);
-
-                    if !prefix.is_empty() {
-                        styled_s.append_plain(Self::decode_html(&prefix));
-                    }
-
-                    styled_s.append_styled(
-                        format!("\"{}\" ", shorten_url(&link)),
-                        Style::from(get_config_theme().link_text.color),
-                    );
-                    styled_s.append_styled(
-                        format!("[{}]", links.len()),
-                        ColorStyle::new(
-                            PaletteColor::TitlePrimary,
-                            get_config_theme().link_id_bg.color,
-                        ),
-                    );
-                    links.push(link);
-                    continue;
-                }
-            }
-        }
-        if !s.is_empty() {
-            styled_s.append_plain(Self::decode_html(&s));
-        }
-        (styled_s, links)
-    }
-
-    /// Parse comments recursively into readable texts with styles and colors
-    fn parse_comments(
-        comments: &[client::Comment],
-        height: usize,
-        top_comment_id: u32,
-    ) -> Vec<Comment> {
-        let paragraph_re = Regex::new(r"<p>(?s)(?P<paragraph>.*?)</p>").unwrap();
-        let italic_re = Regex::new(r"<i>(?s)(?P<text>.+?)</i>").unwrap();
-        let code_re = Regex::new(r"<pre><code>(?s)(?P<code>.+?)[\n]*</code></pre>").unwrap();
-        let link_re = Regex::new(r#"<a\s+?href="(?P<link>.+?)"(?s).+?</a>"#).unwrap();
-
-        comments
-            .iter()
-            .flat_map(|comment| {
-                let top_comment_id = if height == 0 {
-                    comment.id
-                } else {
-                    top_comment_id
-                };
-                let mut subcomments =
-                    Self::parse_comments(&comment.children, height + 1, top_comment_id);
-                let mut comment_string = StyledString::styled(
-                    format!(
-                        "{} {} ago\n",
-                        comment.author,
-                        get_elapsed_time_as_text(comment.time),
-                    ),
-                    PaletteColor::Secondary,
-                );
-
-                let (comment_content, links) = Self::parse_single_comment(
-                    &comment.text,
-                    &paragraph_re,
-                    &italic_re,
-                    &code_re,
-                    &link_re,
-                );
-                comment_string.append(comment_content);
-
-                // minimized_comment is used to display collapsed comment
-                let minimized_comment_string = StyledString::styled(
-                    format!(
-                        "{} {} ago ({} more)",
-                        comment.author,
-                        get_elapsed_time_as_text(comment.time),
-                        subcomments.len() + 1,
-                    ),
-                    PaletteColor::Secondary,
-                );
-
-                subcomments.insert(
-                    0,
-                    Comment::new(
-                        top_comment_id,
-                        comment.id,
-                        comment_string,
-                        minimized_comment_string,
-                        height,
-                        links,
-                    ),
-                );
-                subcomments
-            })
-            .collect()
     }
 
     /// Return the `id` of the first (`direction` dependent and starting but not including `start_id`)
